requirements.txt
django~=5.1.3
django-ckeditor~=6.6.0
pillow~=10.0.1
channels~=4.2.0
asgiref~=3.8.1
channels-redis~=4.1.0  # WebSocket 消息队列
django-crispy-forms~=2.1  # 表单美化
crispy-bootstrap4~=2022.1  # Bootstrap 4 支持
redis~=5.0.1  # Redis 客户端
python-dotenv~=1.0.0  # 读取环境变量
gunicorn~=20.1.0  # 生产环境服务器